Continued refactoring code
