+++ src/main/java/com/github/captnsisko/left4discord/commands/ChatCommand.java
@@ -1,9 +1,9 @@
-package com.github.captnsisko.left4discord.Commands;
+package com.github.captnsisko.left4discord.commands;
 
 import java.util.Timer;
 import java.util.TimerTask;
 
-import com.github.captnsisko.left4discord.Util.Constants;
+import com.github.captnsisko.left4discord.util.Constants;
 
 import org.javacord.api.entity.message.Message;
 import org.javacord.api.entity.message.MessageAttachment;
