+++ src/main/java/com/github/captnsisko/left4discord/commands/TriviaCommand.java
@@ -1,11 +1,11 @@
-package com.github.captnsisko.left4discord.Commands;
+package com.github.captnsisko.left4discord.commands;
 
 import java.awt.Color;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.github.captnsisko.left4discord.TriviaTask;
-import com.github.captnsisko.left4discord.Util.Constants;
+import com.github.captnsisko.left4discord.tasks.TriviaTask;
+import com.github.captnsisko.left4discord.util.Constants;
 
 import org.javacord.api.entity.message.Message;
 import org.javacord.api.entity.message.embed.EmbedBuilder;
